pyproject.toml
[project]
name = "ai-fashion-house"
version = "0.1.8"
description = "AI Fashion House, A multi agent adk application for fashion design"
readme = "README.md"
authors = [
    { name = "haruiz", email = "henryruiz22@gmail.com" },
    { name = "margaretmz", email = "margaretmz@gmail.com"}
]
requires-python = ">=3.11.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.13",
    "db-dtypes>=1.4.3",
    "deprecated>=1.2.18",
    "google-adk>=1.3.0",
    "google-cloud-bigquery>=3.34.0",
    "google-cloud-bigquery-connection>=1.18.3",
    "httpx>=0.28.1",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "rich>=14.0.0",
    "typer>=0.16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pdm.scripts]
ai-fashion-house = "ai_fashion_house.cli:app"

[project.scripts]
ai-fashion-house = "ai_fashion_house.cli:app"



[tool.hatch.build]
exclude = [
    "tests/",
    "docs/",
    "examples/",
    "assets/",
    "keys/",
    ".git/",
    ".github/",
    "scripts/",
    "configs/",
    "README.md",
    ".env",
    "config.json",
    "*/**/.sh",
    "Makefile",
    "cloudbuild.yaml",
    "*/**/Dockerfile",
    "Dockerfile",
    "requirements.txt",
    "docker/",
    "*.sh",
    "*.yml",
    "analysis/",
    "data/",
    "images/",
    "ui/",
    "web.ps1",
]

